12984 FIX Oracle RMAN Backup Monitoring in Dataguard Environments with Archivelog on Standby Side
In Oracle Dataguard environments where the archivelog only activated
on the standby side, the services of the check plugin <tt>oracle_rman</tt>,
which monitors the status of RMAN backups, were never discovered because
the corresponding data was missing from the output of the Oracle agent plugin.
This has been fixed on Linux, Solaris and AIX for Oracle 12.2+. Note that
you need to re-deploy the Oracle agent plugin on affected hosts.

Closes #353

Change-Id: I3044f3c094e579bdbcbea305fbcedad401b43a3e