Label synchronization can now gather discovered labels of a single host
The rule match analyzer of the ruleset editor needs have the discovered
host labels available locally. In this mode we know the site and host we
want to have the labels for and we want to enforce fetching the labels
of that host. This is the base for that operation mode.

Change-Id: I5d4c1066ef6c7189f3ef9651d2371b77847d6d55