Tighten painter content escaping
In previous Checkmk versions the content returned by a painters render
method was always processed as HTML code, even when only a str object
was returned. This unclean handling of the returned data resulted in
possible injection or XSS attacks.

This change requires multiple painters to be changed in the next steps.
Since the str objects are now escaped, the painter render() methods
that want to render HTML code need to return HTML() objects instead
of str objects.

Change-Id: I767c643501165dbe38a04cf28f677f92ec30736c