changes service name from "Cluster Satus" to "Failover state"
The Cisco ASA supports three modes of redundancy / scalability.
 - Failover, the one this check monitors
 - Cluster, what the current service name suggests
 - VPN load balancing cluster, not relevant in this context

Failover consist always of two identical (hardware/software/licenses)
ASA devices. These two devices form a failover pair where one device is
active (handles the Trafic) and the other device is standby. In case of a
failure of the active unit the standby units takes over the active role
and starts to handle the traffic. Both units are connected over a failover link
and optionaly with a state link. In most configuration the failover link and
the state link share one physical interface. In this mode allmost all functions
of the ASA are supported (I belive except for the local CA server). To distinguish
the devices on device is configured as primary on as secondary unit.

In a ASA cluster there can be up to 16 identical devices (ASA5585, Firepower 4100/9330).
The are all active (handle traffic) at the same time. These devices are connected via one control link.
Over this control link the cluster is configured and state information is
passed between the cluster members. In case of asynchron traffic the packets are
routed to the correct device via the cluster link. In this mode of configuration
are many function of the ASA not supported. Failover is particularly important here.

VPN load balancing cluster is mainly for loadbalancing RA VPN sessions. It can be usesd
stand alone or on top of a failover konfiguration (can not be used with cluster mode).
The devices involved hafe not to be identical, there is no state or configuration
synchronisation.

More details can be found in the ASA configuration guide.
https://www.cisco.com/c/en/us/td/docs/security/asa/asa97/configuration/general/asa-97-general-config/ha-cluster.html

This picture gives an overview of the traffic flow in the ASA cluster
https://www.cisco.com/c/dam/en/us/td/i/300001-400000/330001-340000/333001-334000/333480.jpg