Stricter escaping of table.cell()
The content handed over to table.cell() is now escaped more strictly.
Objects of type str handed over are always escaped. Objects of type HTML
are not escaped and rendered as HTML.

The previous default of not escaping HTML in str objects was causing
several injection issues in the past.

Change-Id: Id53d813ad0edcdee1372de7c751cec104091ff1c