gui: view: Prefer list_title for column and sorting
For example `PainterHostCustomVariable` holds a second drop down for
choosing the host custom attribute to display. `title(cell)` evaluates
to the value chosen in this secondary drop down. `title(dummy_cell)`
displays the default (= first) element of the secondary drop down. Prefer
`list_title(cell)` which should show the title of the container.

Change-Id: I9b302f56f3c92294513aedcf8de70a6f4afc19cc