12481 CSV/JSON export of views use custom title for rows headers
Exported view column had a static name depending on the type of the
columns. With parametrizable columns like for Metric History, which content
can be configured and thus always need a title, that same title is used for
export.

The export title is however modified to be all lowercase and all
non-alphanumeric characters are converted to an underscore.

SUP-5034

Change-Id: I8ef8fee1a3c61d9a3fe3f3a222b2dd9158627f1c