12691 systemd_units: addition of factory default levels for reloading and activating in summary check
In the previous version, time thresholds for reloading and activating systemd services
were not part of the default settings for the systemd summary check. This could lead
to an event where a systemd service remained in an activating state for an undefined amount
of time while the check continues to display an OK state. This werk includes both time levels
in the factory default settings in order to prevent such a scenario.

Change-Id: Ifb1a41d1fd0ed5296ec71e45ef9948f7077cc98d