12482 FIX Customgraphs for down hosts
Customgraphs relied on the active metrics of a host to determine renderable
metrics. When a host is down it was assumed that metrics were not
available, and were not rendered.

This werk uses on the customgraph editor the same processing over the
graphs that service graphs have, afterall performance data resides on the
monitoring site not the monitored host. Thus show the metric data of down
hosts.

This behavior was already adopted for CMK-4432 on version 2.0

Change-Id: I396056f82507cd5e04b8d2ede58e6bbdea5c10c6