12726 FIX FreeBSD Agent: Improve detection of executables needed to create monitoring data
The FreeBSD agent checks for various executables, such as zpool or statgrab, to
create monitoring data. If they are not detected, the corresponding sections are
skipped. This werk improves the detection of these executables, which might result
in the discovery of additional services on monitored FreeBSD hosts.

Note that you first have to update the FreeBSD agent on monitored hosts before
this werk comes into effect.

Closes #361

Change-Id: I699e8de4a4ba995856128f13baf7d6da2c85e74c