12559 FIX LDAP: Fix potential crash "KeyError: force_authuser_webservice"
An exception "KeyError: force_authuser_webservice" could occur on the Users
page of the Setup. This was caused by the removal of the deprecated attribute
"force_authuser_webservice" with Checkmk 2.0 for users which had enabled the
synchronization of this attribute from the LDAP.

The LDAP synchronization is now handling this situation by excluding these
attributes from the synchronization.

Change-Id: Iaba578299d51462b8d3ee7c5b5ee5097a28b975f