add value store to item state
This is the last *major* step in the process
of making the item_state module ready to
replace the value_store module.

With this commit we no longer set/get an
"item state prefix". That concept is replaced
by a proper context manager. By attaching the
context manager to the representation of the
data loaded from disk, we avoid operating of
an uninitialised state in the first place.

Change-Id: I2cacb887092f4e343aae48d415820f5ed1345b0d