12709 FIX agent_proxmox_ve: Crash with KeyError('started_time') when corresponding backup job failed
In case a backup job failed, unfinished VM backup logs do not contain `started_time`, which would
result in unhandled KeyError('started_time') when sorting backup entries.
This change will make `agent_proxmox_ve` use the time the job failed at as `started_time` for
each VM in order to be able to assign a failed backup to a certain point in time.

Change-Id: Ia2486f074ddde49cf62f346d5d09ddcb3d579699