12487 Extend reports user permissions to allow edit of report elements from other users
Users in a role with permission "Edit foreign reports"(only admins by
default) could access and edit the properties of reports configured by
other users, yet were not able to edit the contents receiving the error
"The requested report does not exist".

With this werk finding, editing and generating reports with configurations
from other users is possible given the correct permissions.

This feature is not for collaborative editing, but so that admins and other
privileged users can overwrite configurations of other users.

SUP-5193

Change-Id: Ic6bf456a11105d2de60e20d3ec721006cb5fcd57