Improve consistency of some ListColumn::getValue()
The code is actually identical in several classes but this was
obscured but the use of different variable and function names.

This patch uses a consistent naming scheme in every implementation
and links the classes in a comment.

This shall help us refactor and later get rid of the C&P.

Moreover, this demonstrates that the list columns work slightly
differently from the non-list columns and that we should not
replace ListColumn::getValue directly with a callback.

This is a half-mechanical change.

CMK-7076

Change-Id: I007fa583621ada45fe080f0343c315f10826c4f1