12711 FIX ps_check: crash on non-matching enforced service 'State and count of processes'
Configuring 'State and count of processes' as enforced service with per-process details set to
'text' crashes with unhandled "ValueError: 'notice' must be non-empty str or None, got" if the
specified process could not be found.
This change avoids this exception by not trying to render a process list without processes found.

Change-Id: If4f5d8a8527e35d82855e8ba004d4bf55dc2efa4