12730 FIX <tt>mk_zypper</tt>: Increase timeout from 10 to 50 s
The agent plugin <tt>mk_zypper</tt>, which collects data about updates
on SUSE Linux hosts, had an built-in timeout of 10 s. This was too
short, since <tt>zypper refresh</tt> can run for 20 s or more.
This werk increases the timeout to 50 s.

Note that this plugin is normally executed asynchronously and in large
time intervals, so increasing this timeout is not problematic. For this
werk to come into effect, you need to update the agent on affected
hosts.

Closes #364

Change-Id: I8a36b45e26c2333482b2908711f1af211e55de7e