Pass `Mode` to `FileCache.{read,write}`
This lets us reduce feature envy between the fetcher (particularly
in `Fetcher._fetch()`) as well as reduce code duplication in the
agent fetchers.  This is also more OO in that the we move the
responsibility of doing something with the data from the caller to
the callee.  Indeed, the caller does not need to know or care
whether the callee does something with it.

CMK-7752

Change-Id: Ic9fec72f6eae8a7fd13ca7a9f53a447d2f584b10