12735 FIX SMART agent plugin: Exclude devices which can read CDs
The SMART agent plugin now attempts to exclude devices which
can read CDs. Such devices do not support SMART and not
excluding them can lead to hanging <tt>smartctls</tt>
processes. The plugin attempts to exclude such devices based
on the output of <tt>udevadm</tt>, if this command is found.

Note that you need to update the agents on affected linux hosts
for this werk to come into effect.

Closes #365

Change-Id: I1d451271fcec7933766cf7883ddd3f3775c468db