12561 FIX Fix KeyError (site_id) crash on host properties page (2.0.0p2 regression)
When opening the host properties page in the setup an exception KeyError
("site_id") could appear in distributed setups. This issue was triggered when a
folder has a site attribute configured which does not have a site specific
global setting defined. This situation is now cleanly handled to prevent this
exception.

Change-Id: I0bf2a842a1ae67cade02694897dd197df0cd919f