12562 FIX Fix RuntimeError when opening a folder configuration (2.0.0p3 regression)
When opening a folder in setup parts of the Checkmk configuration is loaded in
background. While loading this configuration also old configuration formats are
transformed for a seamless migration of old configuration settings. One of these
transformations of old livestatus proxy configuration could fail with an error
showing the message: "RuntimeError (dictionary changed size during iteration)".

This issue was present in previous versions but was not triggered on folder
configuration pages, but only on global settings or site management pages.
This fix solves the issues for all of those places.

Change-Id: Ia41cfbf31d3ae265e3aadb8e3c5526d41538c3ef