12785 Fix: gui: "A negative value is not valid."
Chromium based browsers showed an error message in the developer tools:

Error: <rect> attribute width: A negative value is not valid.

This error was provoked by a broken windowing function in the python
implementation and is now fixed.

Furthermore the calculation of the windowing function was adapted:
Before the first and last Bar was not a full sized window based on the
current time. Now all time windows have the same duration.

Change-Id: I283e651ba22f07de7119150aa419ad47019f7bab