12565 FIX Prevent log entry in web.log on each page access caused by 2.0 cookies
Users which have authentication cookies issued by a Checkmk 2.0 site caused a
log entry on each page access in the web.log. The log message looked like this:
<tt>Exception while checking cookie auth_[site]: ... ValueError: could not
convert string to float:</tt>.

Since the authentication cookies of both versions are not compatible at all,
we now silently skip processing the cookies of Checkmk 2.0 or newer.

Change-Id: Ifb7f9c967756894a1091e506af057da9a9d894eb