12758 FIX notification replay not working and errors in the cmc.log
If a "Recent notification (for analysis)" in Setup -> Events -> Notifications
was replayed an error message was shown in the User interface. Furthermore,
the error message:

invalid signal byte ASCII 76, expected '*'
read 76 further bytes: 'oading autochecks from /omd/sites/stable/var/check_mk/autochecks/stable.mk

was logged in the cmc.log if a notification was sent. This is fixed.

The reason was that when the labels of a service in complete_raw_context
were loaded a message was printed to the console in autochecks.py. This
interfered with the notification communication of the CMC. The console
logging is now replaced with the logger module. Messages are now logged
in the notify.log. When the command line (e.g. cmk -vII HOST) is used
messages are printed to stdout as before.

CMK-6123

Change-Id: Ibd06516ea71f6e3005ce76d4fc59937364b29234