12343 FIX fix missing __id in folders after upgrade from 1.6.0
When upgrading from 1.6.0 to 2.0.0 until 2.0.0p3, only the
root-folder got assigned a __id key which is required for the
folder endpoints of the REST API. Access to other folders with
the REST API would result in a HTTP 500 error with an error message.

During the upgrade to 2.0.0p4, these folders will be fixed.

No further action (apart from upgrading) is necessary.

Change-Id: Icd7ed54a2cfbba7131ba300d8c2002b2a0286401