Customgraphs editor use only select2
In order to use the modern autocomplete dropdowns for input in host,
service, and metric the old valuespec is completely removed.

Functionality is good enough
- There is no automatic refresh on any form action. This is even a
  feature request: SUP-2485, SUP-5399, SUP-5949
- Added an apply button to save changes
- No need for metric & scalar selector ajax pages. The autocompleter
  covers all
- Because no form & page refresh, can drop the JS part that did that,
  and caused some bugfixing headaches when styles were changes. All
  dropdowns use id instead of name like text_input

In the single info cases and custom graphs editor, where only a specific
hostname, service, and metric make sense enforce that constrain instead
of allowing user to write anything.

 Filters have direct html render, enforce the exact match op=="=" for
 only what is autocompleted.

Change-Id: I685fc48418ae3f743068cbdbd27f2ea3839f5c5c