12344 FIX Status GUI: prevent duplicated export links to REST API
In host- and service-views, when multiple sites were configured,
multiple (dependent on the number of sites configured) identical
links to the REST API were shown in the "Export" drop down menu.

This werk fixes this behaviour. Only unique links will be shown
from now on.

No further action is necessary.

Change-Id: Ie7b241d68bd859e8e3ce37efc026e14767fd3211