12511 FIX Any custom Perf-O-Meter destroyed all Perf-O-Meters
Users can implement their own Perf-O-Meters in
<i>local/share/check_mk/web/plugins/perfometer</i>.
Whenever a user did that in Checkmk 2.0.0 to 2.0.0p4,
<b>all</b> Perf-O-Meters vanished. Instead the error
message "Exception: _required_names" was displayed.

Internal note:

When loading the shipped 'metric' plugins in cmk.gui.metrics
that also includes the shipped perf-o-meters.
The *local* perf-o-meters are in a differently named
directory, and loaded from cmk.gui.views.py (which
cannot be changed easily).

In that case the the postprocessing is ommitted - so we
do the calculations on demand to avoid the
"KeyError: '_required_names'".

Change-Id: I9ff7544858fab978d5c3224d9dfe1d300e5fc9bb