12828 FIX Distributed setup: Fix parallel activations breaking remote sites
Previous 2.0.0 versions could break the Checkmk configuration on remote sites
when two users tried to activate changes for a single remote site at the same
time.

This was a race condition between both processes. While the first activation
performed the synchronization to the remote site, the second activation could
modify the files that are currently being synchronized by the first process.
This could lead to incomplete configurations on the remote site.

Since the first activation did not know anything about this inconsistency, it
tried to activate the incomplete configuration on the remote site after the
sychronization which could fail with different error messages depending on
which files were not transfered.

This issue only affected the configuration of remote sites. No configuration
was lost on the central site.

If you want to prevent this issue with previous 2.0.0 releases, you need to
ensure that you don't start two activations for a site at the same time.

Change-Id: Ifb319e9ee3507a1d7e396b0aac486af60b256b2e