12876 FIX Checking of output phases of UPSs and PDUs: Various fixes
<ul>
<li>The rule "Parameters for output phases of UPSs and PDUs" allows for
configuring levels on the load of the output phases. However, the configured
levels had no actual impact on the monitoring behavior.</li>
<li>Any levels with a warning value of exactly 0 configured in the rule above
had no impact on the monitoring behavior.</li>
<li>Values excatly equal to upper thresholds configured in the rule above were
still considered as below this threshold. This is inconsistent with the general
behavior of Checkmk and has been fixed.</li>
</ul>

SUP-6295

Change-Id: I08eb22e12075a4cdda6e5417496bed6ca2c037c0