Add information to failing Node builds
Node-SASS takes a long time and g++ to be built.
In order to be able to execute the integration Test fast,
we use a pre-built node-sass module from GuitHub:
https://github.com/sass/node-sass/releases

If the build fails, most likely the module we need is not
cached in our Artifact storage. Please upload the corresponding
package.

Change-Id: I39b076849164f751a886712bc49b9477ec3af847