[Draft] Bring back agent backery colored update button
Page menus do support css styling on the div tag. Yet the link inside
has strict CSS and overrides. The important flag is placed to win again.
That is a bad practice, yet is the same strategy used on:

td.count.state a

for the view tables. We need a better strategy on styling.

Change-Id: I0fc27d1ab968b2de30d15f047bdbf1a15d4451e9