12810 FIX check_mk_agent.solaris: fix error "integer expression expected"
The check_mk_agent for solaris reported the error "integer expression expected"
while executing the "<<<job>>>" section.
The cause of this error is that solaris does not support "su -s".
This bugfix removes the parameter "-s" from the su command.
Therefore statistics for jobs of users that have no login shell (e.g. www-data)
are not supported for solaris systems.
In order to apply this bugfix you have to redeploy the agent.

Change-Id: I54f401b0cd7fd3d006405c68ee3dce3b18b2519b