Fix Dashboards show context info in title
Context filters can be configured on the dashboard settings. Those
filters were applied yet did not appear on the title. Only after using
the sidebar filter actions would they be loaded on the url and then
applied to the title.

The context is already merged on the dashboard, and that should be the
only source for action. No last instant query to the uri vars. Views
load at some point the merged context to the uri, that should be
refactored away. Dashboard with context filters is new and that has not
been incorporated( good ).

This commit uses the merge context as the source of information for the
context sensitive title. Hopefully in later refactor can the filters
stop peaking to the uri vars for action.

SUP-6284

Change-Id: I2274f56491d0173d69a5bc386a54f13f2dd179ac