Revert "12878 FIX <tt>systemd_units</tt>: Handle "●" as marker for broken units correctly"
This commit uncovers a lot of issues with previously failed systemd
units. It also shows a lot of check-mk-agent.socket failures. Most of
these uncovered "issues" are mostly not actionable problems right now
because hey shows failures states that happened in the past and have
been fixed in the meantime (just missed to acknowledge the failure
state).

Even if this change is correct, we first need to figure out how to we
and users can deal with these failures before publishing it as part of a
patch release.

This reverts commit 151ba19c5c51a70c8c64a4f1ea40aab85b0f846d.

Change-Id: I9dae710de208d8f254a1391c470c2f4cbce607aa