12993 FIX Filter Multiple WATO folder can now be saved
The Multiple WATO folder filter could be used from the Dashboards and Views
filter menu, yet it could not be saved correspondinly as a context filter, when
editing the view or dashboard. This is now fixed.

Use dual list choice because is a manageable valuespec. The problem with
ListOf is that when the valuespec is to complicated to keep track of it
without the valuespec and it can not be simply recovered the indexing
gets lost among other issues.

SUP-5806
SUP-6263

Change-Id: I40315510bc7755c947945e5ce08317ee3330df15