Make source list creation independent of Mode
The creation of the list of sources is indeed indendant of the Mode.
This patch reflects this fact in the type system.

The resulting API for the `Source` is now consistent with
the Fetchers API, cf. Fetcher.fetch(Mode) => Source.fetch(Mode) and
Summarizer.summarize(HostConfig, *, Mode) =>
Source.summarize(HostConfig, *, Mode).

This is a mechanical change performed with the help of mypy.

Change-Id: I2c35874ba506eaca937eed16df1d6c9bf2f1a80f