12684 FIX agent_kubernetes: inaccurate timestamp conversion
Seen while working on CMK-7644

time.mktime is not the correct function to use to convert timestamp
strings to seconds since epoch, because it always assumes the input to
be in local time. This is even though the datetime object is time zone
aware. See:

https://docs.python.domainunion.de/3/library/time.html#time.mktime

This lead to wrong timestamps being calculated if the remote server was
set to a different time zone than the localhost.

Also, the current time zone of the localhost is irrelevant for
calculating timestamps from a remote server.

Change-Id: I6ecc86bacef0c6ee02c784f8382087113a926d00