Enable eval-used checker
We still have some eval calls left that need some larger rework. Some of
them are only in place for compatibility with older user files. But to
have them in place with the suppressions is better than not having
eval-used enabled because it prevents new call sites from being added.

Change-Id: I54262c7582259ce8ccca2030fc4d230bceaa4226