12879 FIX Host tags in rule conditions: show correct group
Multiple tags in different tag groups can have the same id. If
this was the case, the column "Conditions" in the views showing
the configured rules of one rule set might have shown the wrong
tag group for such tags (depending on the actual rule
configurations). This has been fixed.

SUP-4741

Change-Id: If12447f2fa370588ccde89f0d36d9cef5fde877e