Extract timeout_manager from html class
The timeout handling was already separated in the TimeoutManager but
still a helper of the global html object. It has now been removed from
the html class and is optionally handed over to the RequestContext now.

This also reduces the side effect of the timeout being enabled when
constructing a html object.

Change-Id: I59e298ad242ef42c9bd77a75600996cf3d47dee1