12881 New special agent to monitor Datadog
This werk introduces a new special agent which enables the monitoring
of Datadog monitors and Datadog events in Checkmk. To collects its
data, the agent connects to the Datadog API.

Datadog monitors are monitored as services of the host on which the
special agent is executed. One monitor corresponds to one service.
Datadog events are forwarded to the local event console of the
Checkmk site.

SUP-3069

Change-Id: I43e1fcf96bad8934f17517b56cca22c60f6323de