12764 FIX agent_prometheus: don't crash if the cAdvisor is used without a namespace prefix
This werk only affects you if you use the scrape target cAdvisor with the option
"Entity level used to create Checkmk piggyback hosts" set to "Pod". If "Prepend
namespace prefix for hosts" was not selected, the agent crashed with an error
message. To be consistent with the remaining entity levels and the scrape target
"Kube-state-metrics" the value is now mandatory.

Change-Id: If87670b6f6242f8b6f7aaa47385b166de70bd53a