Allow FileCache usage for CHECKING
Regular, non-cluster cache reading is disabled by `max_age == 0` anyway,
we do not need to handle this otherwise.  Now, cluster data may actually
come from cache provided it is not older than 90 s.

The condition in `cache_read(Mode)` is now identical for SNMP and Agent
fetchers and we may therefore simplify the code in another commit.

This commit adds further logging output in order to make it possible
to understand which path is taken, and why.

CMK-7752

Change-Id: I46717ac29474aafa6111bff505078901d04d7e65