12882 FIX Host tags: Take tag groups into account when matching against rule conditions
Up to now, the matching of host tags against rule conditions was based solely on
on the tag ids. The ids of the tag groups were not taken into account. However,
tag ids are not required to be unique across different tag groups. Therefore,
if the same tag ids occur in multiple tag groups, ignoring the group ids can lead
to wrong results.

This werk fixes this issue. Tag group ids are now taken into account when checking
if a rule applies to a host, both for normal monitoring rules and BI aggregations.

SUP-5957

Change-Id: I13f174f924550ec31fbe6dc528973c652770a73f