Working filter creation from the side filter menu
Discard on page_views the concept of single infos. Filters take the dict
of html vars an should thus just use that construct. No need for the
single string option.

The refresh is not working, no persistence yet. Taking the filter as
overwrite, yet when not active return to context merge.

Refresh view including _active urlvar is needed because now the context
is not "merged" upon spec & vars but taken directly out of the filter
menu, which starts from the spec and is changed by user. This makes it
completely user controlled.

Bug of werk 11759 does not come up because filter value is str_mandatory
and now work is directly on the context dict.

Change-Id: I44d0c19e4f917d225f1013601c58bb51ef52f941