12790 FIX kubernetes: label empty endpoint_info
Before this werk endpoint hosts where only labeled if there was at least
some information about the endpoint available. This is problematic as
the host is created anyway even if there is no information available. In
order to prevent the problem described in werk #12782 the host label
<tt>cmk/kubernetes:yes</tt> has to be available, so we need to label
the host even if there is no information available.

Change-Id: I9be1db6c1bddfc825a462899c971f7929c1d9728