Compose output/render operation on some ListColumns
CommentColumn, DowntimeColumn, ServiceGroupMembersColumn, and
ServiceListColumn only differ on getEntries() and the implementation
of the `output()` member function.  `getEntries()` is analogous to the
`getValues()` functions of our `Column::Callback` types and we shall put
the implementation into a lambda.

The `output()` function needs some more information so that we make a
new class from it.

CMK-7076

Change-Id: Ia6d206a415b69ab4af1ae87bd071bb9f71f943fe